backend/requirements.txt
# FastAPI and Server (Updated versions)
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# Google Gemini API
google-genai>=1.21.0

# LangChain for RAG Pipeline (Updated versions)
langchain>=0.1.10
langchain-google-genai>=1.0.0
langchain-community>=0.0.20

# Core Dependencies
python-dotenv>=1.0.0
pydantic>=2.6.0

# Vector Database
chromadb>=0.4.20
# pinecone-client>=3.0.0  # Uncomment for production

# Code Parsing (Updated versions)
tree-sitter>=0.20.4
tree-sitter-python>=0.21.0
tree-sitter-javascript>=0.20.1
tree-sitter-typescript>=0.20.2

# Embedding Models
sentence-transformers>=2.4.0

# Additional Utilities
requests>=2.31.0
aiofiles>=23.2.1

# Development and Testing
pytest>=7.4.3
httpx>=0.27.0 